+++ packages/starter/src/utils/project.spec.ts
@@ -229,7 +229,6 @@ describe('project utils', () => {
         license: 'ISC',
         main: 'index.js',
         name: 'my-app',
-        scripts: { test: 'echo "Error: no test specified" && exit 1' },
         version: '1.0.0',
       });
     });