+++ packages/starter/src/utils/project.ts
@@ -124,9 +124,6 @@ export const createPackageJSON = (projectName: string): PackageJsonObject => ({
   license: 'ISC',
   main: 'index.js',
   name: projectName,
-  scripts: {
-    test: 'echo "Error: no test specified" && exit 1',
-  },
   version: '1.0.0',
 });